Update minimum node version to 10.15.3 (#103)
* update minimum node version to 10.15update dependencies, minor code changes

* Update README.md for changes to nodejs v10.15.3

* resolve conflicts
