+++ gulp/tasks/02-custom-js.js
@@ -14,12 +14,12 @@ const sourcemaps = require('gulp-sourcemaps');
 
 let buildParams = config.buildParams;
 
-gulp.task('watch-js', ['select-view'], () => {
+gulp.task('watch-js', gulp.series('select-view', () => {
     gulp.watch([`${buildParams.viewJsDir()}/**/*.js`,'!'+buildParams.customPath()],['custom-js']);
-});
+}));
 
 
-gulp.task('custom-js', ['select-view', 'custom-html-templates'],() => {
+gulp.task('custom-js', gulp.series('select-view', 'custom-html-templates',() => {
    if(config.getBrowserify()) {
        return buildByBrowserify();
    }
@@ -27,7 +27,7 @@ gulp.task('custom-js', ['select-view', 'custom-html-templates'],() => {
        return buildByConcatination();
    }
 
-});
+}));
 
 function getBrowserifyBabelPlugins() {
     return [
